refactor: finish sdk-free runtime integration
